backend/requirements.txt
argon2-cffi==25.1.0     
argon2-cffi-bindings==25.1.0 
cloudinary==1.44.1
cryptography==46.0.2
fastapi==0.116.1
google-generativeai==0.8.5
passlib==1.7.4
pip==25.2
pipreqs==0.4.13
pydantic==2.11.9
pydantic-settings==2.11.0
PyMySQL==1.1.2
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
requests==2.32.4
SQLAlchemy==2.0.43
typing_extensions==4.14.1
typing-inspection==0.4.1
uvicorn==0.35.0
